LayoutTests/fast/images/text-recognition/avoid-image-overlay-in-small-image-expected.txt


PASS internals.shadowRoot(image)?.getElementById("div#image-overlay") is undefined.
PASS successfullyParsed is true

TEST COMPLETE
